+++ Consoul/Consoul.csproj
@@ -9,10 +9,10 @@
     <PackageTags>Console</PackageTags>
     <Authors>tbm0115</Authors>
     <Company />
-    <Version>1.4.7</Version>
+    <Version>1.4.8</Version>
     <PackageReleaseNotes>Converted views and view options to run asynchronously to avoid repeated Stack Overflow issues.</PackageReleaseNotes>
-    <AssemblyVersion>1.4.7.0</AssemblyVersion>
-    <FileVersion>1.4.7.0</FileVersion>
+    <AssemblyVersion>1.4.8.0</AssemblyVersion>
+    <FileVersion>1.4.8.0</FileVersion>
     <PackageIcon>Consoul_Small.png</PackageIcon>
     <PackageIconUrl />
   </PropertyGroup>
